refactor generate file into a module
